+++ backend/canisters/community/impl/src/lib.rs
@@ -569,7 +569,7 @@ impl RuntimeState {
         let member = self.data.members.get_verified_member(caller)?;
 
         match member.user_type {
-            UserType::User => Ok(Caller::User(member.user_id)),
+            UserType::User => Ok(Caller::User(member.user())),
             UserType::Bot => Ok(Caller::Bot(member.user_id)),
             UserType::OcControlledBot => Ok(Caller::OCBot(member.user_id)),
             UserType::BotV2 | UserType::Webhook => Err(OCErrorCode::InitiatorNotFound),